+++ .gdb_history
@@ -1,38 +1,3 @@
-p trapframe
-p trap_frame
-p *trap_frame
-b 0x00000028
-b *0x00000028
-c
-b *0x00000028
-c
-b *0x38
-c
-b switch_to_user
-b switch_to_usermode
-b switch_to_user_mode
-c
-b 152
-b 153
-c
-p task->sp
-p task->elr
-p task->utask->sp
-p task->utask->elr
-p 0x0
-p x/10i 0
-x/10i 0
-si
-c
-b 0x38
-b *0x38
-c
-b pgd
-c
-p task->pgd
-si
-b 0x38
-c
 b e
 c'
 c
@@ -254,3 +219,38 @@ p *(task + 1)
 p *(task + 2)
 p (task + 2)
 p *(task + 2)
+si
+b *0x3e4
+c
+p [$sp, 28]
+p $sp + 28
+p *($sp + 28)
+p *($sp + 28)
+p *0xffffffffdfec
+c
+b switch_to
+c
+p $x2
+c
+p $x2
+c
+p $x2
+si
+p $lr
+si
+b *0x3e4
+c
+b 0x0
+c
+b *0x0
+c
+p *0xfffffffff000
+p *0xfffffffff004
+p *0xfffffffff000
+p *0xffffffffa000
+p *0xfffffffff000
+si
+si
+p $lr
+si
+p $lr
