+++ src/libsyscall.c
@@ -1,7 +1,9 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#include "user_lib.h"
+#include "libsyscall.h"
+#include "syscall.h"
+#include "task.h"
 size_t sys_uart_write(const char buf[], size_t size) {
     asm volatile("mov x8, #0");
     asm volatile("svc 1");
@@ -19,12 +21,12 @@ void sys_exec(void (*func)()) {
     asm volatile("svc 1");
 }
 
-/* int sys_fork() { */
-/* asm volatile("mov x8, #3"); */
-/* asm volatile("svc 1"); */
-/* struct task_t* task = get_current(); */
-/* return task->utask.fork_id; */
-/* } */
+int sys_fork() {
+    asm volatile("mov x8, #3");
+    asm volatile("svc 1");
+    struct task_t* task = get_current();
+    return task->utask.fork_id;
+}
 
 void sys_exit(int status) {
     asm volatile("mov x8, #4");
