+++ src/task.c
@@ -6,10 +6,6 @@
 #include "utils.h"
 #define current get_current();
 
-extern void switch_to(struct task_t* prev, struct task_t* next,
-                      uint64_t nextfunc, uint64_t spsr, uint64_t* pgd);
-extern void user_context(uint64_t sp, uint64_t func);
-
 struct task_t task_pool[64];
 struct queue_element_t queue_elements[QUEUE_ELE_SIZE];
 int queue_elements_now;
@@ -142,111 +138,9 @@ void schedule() {
     context_switch(task);
 }
 
-struct utask_t* get_current_utask() {
-    uint64_t res;
-    asm volatile("mrs %0, tpidr_el0" : "=r"(res));
-    struct utask_t* res_task = (struct utask_t*)res;
-
-    return res_task;
-}
-
 void switch_to_user_mode() {
     struct task_t* task = get_current();
     uint64_t sp = task->utask.sp;
     uint64_t func = task->utask.elr;
     user_context(sp, func);
 }
-
-void do_exec(uint8_t* func, int size) {
-    struct task_t* task = get_current();
-
-    struct page_t* user_page = page_alloc();
-    task->pages[task->pages_now++] = user_page->id;
-    asm volatile("b:");
-    for (int i = 0; i < size; i++) {
-        *((uint8_t*)user_page->content + i) = *(func + i);
-    }
-    page_mapping(task, user_page);
-    task->utask.elr = 0;
-
-    struct page_t* stack_page = page_alloc();
-    task->pages[task->pages_now++] = stack_page->id;
-    stack_mapping(task, stack_page);
-    task->utask.sp = 0x0000ffffffffe000;
-
-    asm volatile("pgd:");
-    uint64_t utask_addr = (uint64_t)&task->utask;
-    move_ttbr(task->pgd);
-
-    asm volatile("mov     x6, %0" : "=r"(utask_addr));
-    asm volatile("msr     tpidr_el0, x6");
-    asm volatile("ldr x2, =switch_to_user_mode");
-    asm volatile("msr     elr_el1, x2");
-    asm volatile("bl      set_aux");
-    asm volatile("eret");
-}
-
-void do_fork(uint64_t elr) {
-    struct task_t *task = get_current(), *child_task = 0;
-    uint64_t sp_el0;
-    asm volatile("mrs %0, sp_el0" : "=r"(sp_el0));
-    task->utask.sp = sp_el0;
-    for (int i = 0; i < 64; i++) {
-        if (task_pool[i].status != ACTIVE) {
-            child_task = &task_pool[i];
-            struct page_t* user_page = page_alloc();
-            task_pool[i].pages[task_pool[i].pages_now++] = user_page->id;
-
-            memcpy(user_page->content, pages[task->user_page].content,
-                   4 * 1024);
-
-            page_mapping(&task_pool[i], user_page);
-            task_pool[i].utask.elr = 0;
-
-            struct page_t* stack_page = page_alloc();
-            task_pool[i].pages[task_pool[i].pages_now++] = stack_page->id;
-            stack_mapping(&task_pool[i], stack_page);
-
-            memcpy(stack_page->content, pages[task->stack_page].content,
-                   4 * 1024);
-
-            task_pool[i].id = i;
-            task_pool[i].elr = elr;
-            task_pool[i].time = 0;
-            task_pool[i].pages_now = 0;
-            task_pool[i].spsr = task->spsr;
-            task_pool[i].priority = task->priority;
-            task_pool[i].status = ACTIVE;
-            task_pool[i].utask.elr = task->utask.elr;
-            task_pool[i].utask.sp = task->utask.sp;
-            task_pool[i].sp = task_pool[i].utask.sp;
-            task_pool[i].signal = 0;
-            task_pool[i].utask.fork_id = 0;
-            /* task_pool[i].trap_frame->x0 = 0; */
-            memcpy(&kstack_pool[i - 1] + 1, &kstack_pool[task->id - 1] + 1,
-                   STACK_SIZE * sizeof(char));
-            task->utask.fork_id = task_pool[i].id;
-            task->trap_frame->x0 = task_pool[i].id;
-            break;
-        }
-    }
-    queue_push(&runqueue, child_task);
-}
-
-void kexit(uint64_t status) {
-    do_exit(status);
-    schedule();
-}
-
-void do_exit(uint64_t status) {
-    struct task_t* task = get_current();
-    task->status = ZOMBIE;
-    for (uint64_t i = 0; i < task->pages_now - 1; i++) {
-        page_free(&pages[task->pages[i]]);
-    }
-    print_s("Exited with status code: ");
-    print_i(status);
-    print_s("\n");
-}
-
-void do_kill(uint64_t pid, uint64_t signal) { task_pool[pid].signal = signal; }
