+++ src/trap.c
@@ -2,6 +2,7 @@
 
 #include "io.h"
 #include "page.h"
+#include "syscall.h"
 #include "task.h"
 #include "timer.h"
 #include "trap.h"
@@ -45,38 +46,8 @@ void synchronize_handler(uint64_t esr, uint64_t elr,
     }
     if (iss == 1) {
         uint64_t syscall = trap_frame->x8;
-        char tmp;
-        switch (syscall) {
-            case 0:
-                for (uint64_t i = 0; i < trap_frame->x1; i++) {
-                    uart_send(*((char *)trap_frame->x0 + i));
-                }
-                break;
-            case 1:
-                for (uint64_t i = 0; i < trap_frame->x1; i++) {
-                    *((char *)trap_frame->x0 + i) = uart_getc();
-                }
-                break;
-            case 2:
-                /* do_exec((void (*)())trap_frame->x0); */
-                break;
-            case 3:
-                do_fork(elr);
-                break;
-            case 4:
-                do_exit(trap_frame->x0);
-                asm volatile("ldr x0, =schedule");
-                asm volatile("msr elr_el1, x0");
-                break;
-            case 5:
-                do_kill(trap_frame->x0, trap_frame->x1);
-                break;
-            case 7:
-                tmp = uart_getc();
-                trap_frame->x0 = tmp;
-                break;
-        }
-        asm volatile("be:");
+        syscall_handler(syscall, trap_frame, elr);
+
         uint64_t *pgd = task->pgd;
         move_ttbr(pgd);
     } else if (iss == 2) {
